requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
imagehash==4.3.1
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
